Timeout when trying to connect to the k8s API server
Fixes #76

If we don't specify a timeout on the session.get() call, we get wildly
varying reconnect timeout behaviour as described in the linked GitHub
issue.

Also reduce logging for connection errors from log.exception to
log.warning.  It _is_ a problem if we can't connect, and so we do want
the log to flag that; but if the API server is away for a while, we
don't need the same traceback every time.